Now with a player (and concept art)
I copied over just the PlayerController and GameManager classes from Time Leap. I removed all references to power ups, player 1/player 2, and moving platforms. I commented out references to checkpoints, animations. and the input system we were using. For the input, I changed it for now to the old, built-in Unity system that we were using at the beginning. Once we decide on what we are planning to do with the commented out areas we can remove/reimplement them. I copied over the constants for PlayerController from Time Leap, although the player is a slightly different size.