Added assignment prefab
I made an assignment (enemy) prefab. I pretty much copied some code from PlayerController to make it move like a player. As of right now, its horizontal movement is controlled by the user for testing, and it can't move itself vertically. It doesn't collide with the player or other assignments, but does collide with walls and stuff (thanks to the layer system). Currently, walking into an assignment does nothing, but if we want to do something we will need to add another collider that acts as a trigger.

In the future, it will have some kind of AI to move it left and right but not walk off the edge.