New scene, new camera, more orgaization
I made prefabs from the old scene and opened up a new scene (with not much in it). I re-added the camera following from Time Leap and tweaked the Shoot script so that it works with a perspective camera (that will give us parallax). I also organized our assets into their respective folders.